backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
supabase>=2.10.0
networkx>=3.4
python-dotenv>=1.0.0
pydantic>=2.10.0
websockets>=14.0
httpx>=0.28.0
